clean up example
