+++ examples/sync/src/databases/any.rs
@@ -2,7 +2,7 @@
 
 use crate::{Hasher, Key, Translator, Value};
 use commonware_cryptography::Hasher as CryptoHasher;
-use commonware_storage::{adb::any::fixed, store::operation};
+use commonware_storage::{adb::any::fixed, mmr::hasher::Standard, store::operation};
 use commonware_utils::NZUsize;
 
 /// Database type alias.
@@ -28,47 +28,108 @@ pub fn create_config() -> fixed::Config<Translator> {
     }
 }
 
-/// Create deterministic test operations for demonstration purposes.
-///
-/// This function creates a sequence of Update operations followed by
-/// periodic Commit operations. The operations are deterministic based
-/// on the count and seed parameters.
-pub fn create_test_operations(count: usize, seed: u64) -> Vec<Operation> {
-    let mut operations = Vec::new();
-    let mut hasher = <Hasher as CryptoHasher>::new();
+impl<E> crate::databases::Syncable for Database<E>
+where
+    E: commonware_runtime::Storage + commonware_runtime::Clock + commonware_runtime::Metrics,
+{
+    type Operation = Operation;
 
-    for i in 0..count {
-        let key = {
-            hasher.update(&i.to_be_bytes());
-            hasher.update(&seed.to_be_bytes());
-            hasher.finalize()
-        };
+    fn create_test_operations(count: usize, seed: u64) -> Vec<Self::Operation> {
+        let mut hasher = <Hasher as CryptoHasher>::new();
+        let mut operations = Vec::new();
+        for i in 0..count {
+            let key = {
+                hasher.update(&i.to_be_bytes());
+                hasher.update(&seed.to_be_bytes());
+                hasher.finalize()
+            };
 
-        let value = {
-            hasher.update(&key);
-            hasher.update(b"value");
-            hasher.finalize()
-        };
+            let value = {
+                hasher.update(&key);
+                hasher.update(b"value");
+                hasher.finalize()
+            };
 
-        operations.push(Operation::Update(key, value));
+            operations.push(Operation::Update(key, value));
 
-        if (i + 1) % 10 == 0 {
-            operations.push(Operation::CommitFloor(i as u64 + 1));
+            if (i + 1) % 10 == 0 {
+                operations.push(Operation::CommitFloor(i as u64 + 1));
+            }
         }
+
+        // Always end with a commit
+        operations.push(Operation::CommitFloor(count as u64));
+        operations
     }
 
-    // Always end with a commit
-    operations.push(Operation::CommitFloor(count as u64));
-    operations
+    async fn add_operations(
+        database: &mut Self,
+        operations: Vec<Self::Operation>,
+    ) -> Result<(), commonware_storage::adb::Error> {
+        for operation in operations {
+            match operation {
+                Operation::Update(key, value) => {
+                    database.update(key, value).await?;
+                }
+                Operation::Delete(key) => {
+                    database.delete(key).await?;
+                }
+                Operation::CommitFloor(_) => {
+                    database.commit().await?;
+                }
+            }
+        }
+        Ok(())
+    }
+
+    async fn commit(&mut self) -> Result<(), commonware_storage::adb::Error> {
+        self.commit().await
+    }
+
+    fn root(&self, hasher: &mut Standard<commonware_cryptography::Sha256>) -> Key {
+        self.root(hasher)
+    }
+
+    fn op_count(&self) -> u64 {
+        self.op_count()
+    }
+
+    fn lower_bound_ops(&self) -> u64 {
+        self.inactivity_floor_loc()
+    }
+
+    fn historical_proof(
+        &self,
+        size: u64,
+        start_loc: u64,
+        max_ops: u64,
+    ) -> impl std::future::Future<
+        Output = Result<
+            (
+                commonware_storage::mmr::verification::Proof<Key>,
+                Vec<Self::Operation>,
+            ),
+            commonware_storage::adb::Error,
+        >,
+    > + Send {
+        self.historical_proof(size, start_loc, max_ops)
+    }
+
+    fn database_name() -> &'static str {
+        "Any"
+    }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::databases::Syncable;
+
+    type AnyDb = Database<commonware_runtime::deterministic::Context>;
 
     #[test]
     fn test_create_test_operations() {
-        let ops = create_test_operations(5, 12345);
+        let ops = <AnyDb as Syncable>::create_test_operations(5, 12345);
         assert_eq!(ops.len(), 6); // 5 operations + 1 commit
 
         if let Operation::CommitFloor(loc) = &ops[5] {
@@ -81,12 +142,12 @@ mod tests {
     #[test]
     fn test_deterministic_operations() {
         // Operations should be deterministic based on seed
-        let ops1 = create_test_operations(3, 12345);
-        let ops2 = create_test_operations(3, 12345);
+        let ops1 = <AnyDb as Syncable>::create_test_operations(3, 12345);
+        let ops2 = <AnyDb as Syncable>::create_test_operations(3, 12345);
         assert_eq!(ops1, ops2);
 
         // Different seeds should produce different operations
-        let ops3 = create_test_operations(3, 54321);
+        let ops3 = <AnyDb as Syncable>::create_test_operations(3, 54321);
         assert_ne!(ops1, ops3);
     }
 }
